DPR-482 extracted multiline editor into separate class
