--- cli/src/main/kotlin/uk/gov/justice/digital/cli/command/MultilineParser.kt
@@ -1,4 +1,0 @@
-package uk.gov.justice.digital.cli.command
-
-class MultilineParser {
-}
